feat: implement cjsToEsm.js
