+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,39 @@
+import path from "path";
+import { release, version } from "os";
+import { createServer as createServerHttp } from "http";
+import jsonA from "./files/a.json" assert { type: "json" };
+import jsonB from "./files/b.json" assert { type: "json" };
+import { fileURLToPath } from "url";
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+  unknownObject = jsonA;
+} else {
+  unknownObject = jsonB;
+}
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+const fileNamePath = fileURLToPath(import.meta.url);
+const filePath = path.dirname(fileNamePath);
+console.log(`Path to current file is ${fileNamePath}`);
+console.log(`Path to current directory is ${filePath}`);
+
+const myServer = createServerHttp((_, res) => {
+  res.end("Request accepted");
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log("To terminate it, use Ctrl+C combination");
+});
+
+export { unknownObject, myServer };
